Move into a separate system.
