+++ workers/unity/Packages/com.improbable.gdk.core/Systems/ReactiveComponentAdditionSystem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b2e2e3648f6881842984b3777a7c91c8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
