+++ workers/unity/Packages/com.improbable.gdk.core/Worker/WorkerBase.cs
@@ -83,6 +83,7 @@ namespace Improbable.Gdk.Core
         {
             World.GetOrCreateManager<EntityManager>();
             World.GetOrCreateManager<SpatialOSReceiveSystem>();
+            World.GetOrCreateManager<ReactiveComponentAdditionSystem>();
             World.GetOrCreateManager<SpatialOSSendSystem>();
             World.GetOrCreateManager<CleanReactiveComponentsSystem>();
         }
